🎨 HAL_MinSerial.* => MinSerial.*
